+++ package.json
@@ -31,6 +31,7 @@
     "good-console": "^5.2.0",
     "hapi": "^12.1.0",
     "hapi-auth-jwt2": "^5.3.2",
+    "hapi-sass": "^2.0.0",
     "inert": "^3.2.0",
     "joi": "^7.2.2",
     "jsonwebtoken": "^5.5.4",
